requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
dataclasses-json==0.5.6
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
marshmallow==3.14.1
marshmallow-enum==1.5.1
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.4
pandas==1.3.4
python-binance==1.0.15
python-dateutil==2.8.2
python-ftx==0.0.4
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
ta==0.8.0
tornado==6.1
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
